feat(🌟User): add get mapping to know  referral code user

+++ src/main/java/com/ilevent/ilevent_backend/events/dto/TicketDto.java
@@ -8,7 +8,7 @@ public class TicketDto {
     private Long id;
     private String nameTicket;
     private Integer availableSeats;
-    private Integer priceBeforeDiscount;
+    private Double priceBeforeDiscount;
 
     public static TicketDto fromEntity(Ticket ticket) {
         TicketDto dto = new TicketDto();
